third_party/WebKit/LayoutTests/inspector/console/console-log-linkify-stack-in-errors-expected.txt

CONSOLE MESSAGE: line 5: Error: Some test
    at namedFunction (foob.js:5:17)
    at foob.js:8:1
CONSOLE MESSAGE: line 10: Error: line
break
    at forStack (console-log-linkify-stack-in-errors.html:10:15)
    at console-log-linkify-stack-in-errors.html:13:1
CONSOLE MESSAGE: line 26: TypeError: Cannot read property 'removeChild' of null
    at domError (console-log-linkify-stack-in-errors.html:24:19)
    at console-log-linkify-stack-in-errors.html:30:1
CONSOLE MESSAGE: line 36: Error: some error
CONSOLE MESSAGE: line 42: Error message without stacks http://www.chromium.org/
CONSOLE MESSAGE: line 44: Error valid stack #2
    at http://www.chromium.org/boo.js:40:70
    at foo(http://www.chromium.org/foo.js:10:50)
CONSOLE MESSAGE: line 47: Error valid stack #3
    at http://www.chromium.org/foo.js:40
CONSOLE MESSAGE: line 48: Error: MyError
    at throwError (http://www.chromium.org/foo.js:40)
    at eval (eval at <anonymous> (http://www.chromium.org/foo.js:42:1), <anonymous>:1:1)
    at http://www.chromium.org/foo.js:239
CONSOLE MESSAGE: line 17: ReferenceError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:17:17)
    at stack1 (console-log-linkify-stack-in-errors.html:19:3)
    at console-log-linkify-stack-in-errors.html:52:1
CONSOLE MESSAGE: line 17: EvalError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:17:17)
    at stack1 (console-log-linkify-stack-in-errors.html:19:3)
    at console-log-linkify-stack-in-errors.html:53:1
CONSOLE MESSAGE: line 17: SyntaxError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:17:17)
    at stack1 (console-log-linkify-stack-in-errors.html:19:3)
    at console-log-linkify-stack-in-errors.html:54:1
CONSOLE MESSAGE: line 17: RangeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:17:17)
    at stack1 (console-log-linkify-stack-in-errors.html:19:3)
    at console-log-linkify-stack-in-errors.html:55:1
CONSOLE MESSAGE: line 17: TypeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:17:17)
    at stack1 (console-log-linkify-stack-in-errors.html:19:3)
    at console-log-linkify-stack-in-errors.html:56:1
CONSOLE MESSAGE: line 17: URIError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:17:17)
    at stack1 (console-log-linkify-stack-in-errors.html:19:3)
    at console-log-linkify-stack-in-errors.html:57:1
CONSOLE MESSAGE: line 59: Error broken stack
    at function_name(foob.js foob.js:30:1)
 at foob.js:40:70
CONSOLE MESSAGE: line 60: Error broken stack #2
    at function_name(foob.js:20:30
CONSOLE MESSAGE: line 61: Error broken stack #3
    at function_name(foob:20.js:30   bla
CONSOLE MESSAGE: line 62: Error broken stack #4
    at function_name)foob.js:20:30(
CONSOLE MESSAGE: line 63: Error broken stack #5
    at function_name foob.js:20:30)
CONSOLE MESSAGE: line 64: Error broken stack #6
    at foob.js foob.js:40:70
CONSOLE MESSAGE: line 9: Error
    at Failure.letsFailWithStack (stack-with-sourceMap.js:19:20)
    at <anonymous>:1:1
Test that console.log(new Error().stack) would linkify links in stacks for sourceUrls and sourceMaps Bug 424001.

foob.js:5 Error: Some test
    at namedFunction (foob.js:5)
    at foob.js:8 console-message > source-code > console-message-anchor > devtools-link > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link
console-log-linkify-…ck-in-errors.html:3 Error: line
break
    at forStack (console-log-linkify-…ck-in-errors.html:3)
    at console-log-linkify-…ck-in-errors.html:6 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:19 TypeError: Cannot read property 'removeChild' of null
    at domError (console-log-linkify-…k-in-errors.html:17)
    at console-log-linkify-…k-in-errors.html:23 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:29 Error: some error
    at logError (console-log-linkify-…k-in-errors.html:27)
    at console-log-linkify-…k-in-errors.html:33 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > object-value-error source-code > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:35 Error message without stacks http://www.chromium.org/ console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link
console-log-linkify-…k-in-errors.html:37 Error valid stack #2
    at www.chromium.org/boo.js:40
    at foo(www.chromium.org/foo.js:10) console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link
console-log-linkify-…k-in-errors.html:40 Error valid stack #3
    at www.chromium.org/foo.js:40 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link
console-log-linkify-…k-in-errors.html:41 Error: MyError
    at throwError (www.chromium.org/foo.js:40)
    at eval (eval at <anonymous> (www.chromium.org/foo.js:42), <anonymous>:1:1)
    at www.chromium.org/foo.js:239 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link > devtools-link
console-log-linkify-…k-in-errors.html:10 ReferenceError: valid stack
    at stack2 (console-log-linkify-…k-in-errors.html:10)
    at stack1 (console-log-linkify-…k-in-errors.html:12)
    at console-log-linkify-…k-in-errors.html:45 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:10 EvalError: valid stack
    at stack2 (console-log-linkify-…k-in-errors.html:10)
    at stack1 (console-log-linkify-…k-in-errors.html:12)
    at console-log-linkify-…k-in-errors.html:46 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:10 SyntaxError: valid stack
    at stack2 (console-log-linkify-…k-in-errors.html:10)
    at stack1 (console-log-linkify-…k-in-errors.html:12)
    at console-log-linkify-…k-in-errors.html:47 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:10 RangeError: valid stack
    at stack2 (console-log-linkify-…k-in-errors.html:10)
    at stack1 (console-log-linkify-…k-in-errors.html:12)
    at console-log-linkify-…k-in-errors.html:48 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:10 TypeError: valid stack
    at stack2 (console-log-linkify-…k-in-errors.html:10)
    at stack1 (console-log-linkify-…k-in-errors.html:12)
    at console-log-linkify-…k-in-errors.html:49 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:10 URIError: valid stack
    at stack2 (console-log-linkify-…k-in-errors.html:10)
    at stack1 (console-log-linkify-…k-in-errors.html:12)
    at console-log-linkify-…k-in-errors.html:50 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis > devtools-link > devtools-link-ellipsis
console-log-linkify-…k-in-errors.html:52 Error broken stack
    at function_name(foob.js foob.js:30:1)
 at foob.js:40:70 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text
console-log-linkify-…k-in-errors.html:53 Error broken stack #2
    at function_name(foob.js:20:30 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text
console-log-linkify-…k-in-errors.html:54 Error broken stack #3
    at function_name(foob:20.js:30   bla console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text
console-log-linkify-…k-in-errors.html:55 Error broken stack #4
    at function_name)foob.js:20:30( console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text
console-log-linkify-…k-in-errors.html:56 Error broken stack #5
    at function_name foob.js:20:30) console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text
console-log-linkify-…k-in-errors.html:57 Error broken stack #6
    at foob.js foob.js:40:70 console-message > source-code > console-message-anchor > devtools-link > devtools-link-ellipsis > hidden console-message-badge > hide-badge-title > console-message-text
stack-with-sourceMap.coffee:3 Error
    at Failure.letsFailWithStack (stack-with-sourceMap.coffee:3)
    at window.failure (stack-with-sourceMap.coffee:7)
    at <anonymous>:1:1 console-message > source-code > console-message-anchor > devtools-link > hidden console-message-badge > hide-badge-title > console-message-text > devtools-link > devtools-link
